[COS-955] - Customer map dashboard implementation

+++ packages/server/customer-os-api/entity/dashboard/dashboard.go
@@ -84,6 +84,6 @@ var DashboardCustomerMapStates = []DashboardCustomerMapState{
 type DashboardCustomerMapData struct {
 	OrganizationId     string
 	State              DashboardCustomerMapState
-	Arr                int
+	Arr                float64
 	ContractSignedDate time.Time
 }
